feat: Add generic function for creating arrays with tuples
This commit adds a new generic function, `createArrayWithTuple`, to the `functionGeneric.ts` file. The function takes two parameters of different types and returns an array containing those parameters as a tuple.

The commit also includes examples of using the `createArrayWithTuple` function to create tuples of strings and numbers, as well as tuples of numbers and objects with name and age properties.

Refactor: Update user variable to use IPerson interface in genericTypes.ts