feat: Add makeSound method to Animal class and create instance with sound
This commit adds the `makeSound` method to the `Animal` class in `class.ts`. The `makeSound` method logs the sound the animal makes to the console.

Additionally, an instance of the `Animal` class is created with the name "Dog", species "Canis lupus", and sound "Bark". The `makeSound` method is called on the instance to log the sound "Bark" to the console.