feat: Add id, name, contactNo, and address properties to IUser interface
This commit adds the id, name, contactNo, and address properties to the IUser interface in destructring.ts. The id property is of type number, the name property is an object with firstName, middleName, and lastName properties of type string, the contactNo property is of type string, and the address property is an object with city and country properties of type string.