feat: Update variable types in variableTypes.ts
This commit updates the variable types in the variableTypes.ts file. The myString variable is now commented to indicate that it can be assigned a string value using single quotes or backticks. The isTrue variable is also commented to indicate that it can be assigned a boolean value of true or false.